Make baselayerselector a component
