+++ src/pages/MapPage.tsx
@@ -1,9 +1,10 @@
-// src/components/MapPage.tsx
+// src/pages/MapPage.tsx
 import React, { useState, useRef, useEffect } from 'react';
 import DeckGL from '@deck.gl/react';
 import StaticMap from 'react-map-gl';
 import { LineLayer } from '@deck.gl/layers';
 import 'mapbox-gl/dist/mapbox-gl.css';
+import BaseLayerSelector from '../components/BaseLayerSelector';
 
 const MAPBOX_ACCESS_TOKEN = process.env.REACT_APP_MAPBOX_ACCESS_TOKEN;
 
@@ -25,9 +26,7 @@ const layers = [
 
 const MapPage: React.FC = () => {
   const [mapStyle, setMapStyle] = useState('mapbox://styles/mapbox/light-v9');
-  const [isExpanded, setIsExpanded] = useState(false);
   const [mapHeight, setMapHeight] = useState('100vh');
-  const timeoutRef = useRef<NodeJS.Timeout | null>(null);
   const navbarRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
@@ -46,24 +45,6 @@ const MapPage: React.FC = () => {
     };
   }, []);
 
-  const handleStyleChange = (style: string) => {
-    setMapStyle(style);
-    setIsExpanded(false);
-  };
-
-  const handleMouseEnter = () => {
-    if (timeoutRef.current) {
-      clearTimeout(timeoutRef.current);
-    }
-    setIsExpanded(true);
-  };
-
-  const handleMouseLeave = () => {
-    timeoutRef.current = setTimeout(() => {
-      setIsExpanded(false);
-    }, 5000);
-  };
-
   return (
     <div ref={navbarRef}>
       <div style={{ height: mapHeight }}>
@@ -78,47 +59,10 @@ const MapPage: React.FC = () => {
             style={{ height: '100%' }}
           />
         </DeckGL>
-        <div
-          className="absolute bottom-4 left-4 z-10"
-          onMouseEnter={handleMouseEnter}
-          onMouseLeave={handleMouseLeave}
-        >
-          <div className="relative group">
-            <button className="bg-white p-2 rounded shadow">Base Layer</button>
-            <div className={`absolute left-0 bottom-full mb-2 ${isExpanded ? 'flex' : 'hidden'} flex-col bg-white p-2 rounded shadow transition-opacity duration-500`}>
-              <button
-                className={`p-2 rounded ${mapStyle === 'mapbox://styles/mapbox/light-v9' ? 'bg-gray-300' : 'bg-white'}`}
-                onClick={() => handleStyleChange('mapbox://styles/mapbox/light-v9')}
-              >
-                Light
-              </button>
-              <button
-                className={`p-2 rounded ${mapStyle === 'mapbox://styles/mapbox/dark-v9' ? 'bg-gray-300' : 'bg-white'}`}
-                onClick={() => handleStyleChange('mapbox://styles/mapbox/dark-v9')}
-              >
-                Dark
-              </button>
-              <button
-                className={`p-2 rounded ${mapStyle === 'mapbox://styles/mapbox/streets-v11' ? 'bg-gray-300' : 'bg-white'}`}
-                onClick={() => handleStyleChange('mapbox://styles/mapbox/streets-v11')}
-              >
-                Streets
-              </button>
-              <button
-                className={`p-2 rounded ${mapStyle === 'mapbox://styles/mapbox/outdoors-v11' ? 'bg-gray-300' : 'bg-white'}`}
-                onClick={() => handleStyleChange('mapbox://styles/mapbox/outdoors-v11')}
-              >
-                Outdoors
-              </button>
-              <button
-                className={`p-2 rounded ${mapStyle === 'mapbox://styles/mapbox/satellite-v9' ? 'bg-gray-300' : 'bg-white'}`}
-                onClick={() => handleStyleChange('mapbox://styles/mapbox/satellite-v9')}
-              >
-                Satellite
-              </button>
-            </div>
-          </div>
-        </div>
+        <BaseLayerSelector
+          currentStyle={mapStyle}
+          onStyleChange={(style) => setMapStyle(style)}
+        />
       </div>
     </div>
   );
